multiprocessing+ 1.almost fixed multiprocessing 2.add semaphores via mbx 3.using posix shared memory 4.port atimic load/store 5.vms_sleep 6.reload crlf/lf test files from git

+++ Lib/multiprocessing/heap.py
@@ -83,7 +83,8 @@ else:
                 self.fd, name = tempfile.mkstemp(
                      prefix='pym-%d-'%os.getpid(),
                      dir=self._choose_dir(size))
-                os.unlink(name)
+                if os.sys.platform != 'OpenVMS':
+                    os.unlink(name)
                 util.Finalize(self, os.close, (self.fd,))
                 os.ftruncate(self.fd, size)
             self.buffer = mmap.mmap(self.fd, self.size)
